Split Blameable into CreatedBy and CreatedAndUpdatedBy
CreatedAndUpdatedBy is the renamed Blameable, unchanged in behavior; every
existing consumer is updated to the new name. CreatedBy is a new, smaller
trait with just createdBy/createdOn, for entities that are never edited in
place (e.g. a revision) where "last edited" would be misleading since a
change always produces a new entity instead of updating the existing one.
